Add files via upload
Carro x AWS Hackathon: utilize the three given datasets to come up with an algorithm foretelling and optimizing COE trend for cars in Singapore. Multivariate Regression Analysis are used to create the best possible model for COE price predictions. The model created was evaluated using RMSE (a measure of the model error).